remove the aiaiai-test-defconfigs
It does not work for anymore because we cannot find the correct architecture
(ARCH=) from the defconfigs.

For example, if a patch changes arch/x86/configs/i386_deconfig, the script
finds out that the architecture is 'x86', and then starts verifying the
patch using 'ARCH=x86', which actually translates to 'ARCH=x86_64', which
is incorrect and gives bogus results.

I do not know how to easily fix this and do not have time for a non-trivial
solution, so just removing this script. If later I need it, it is in the git
history.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>