make-kernel: fix --keep-going option
The make-kernel utility has an option "--keep-going" which can be useful
if you want to continue testing past kernel build failures, in order to try
and find as many issues as you can. However, it previously did not work,
as getopt-long format required a parameter, and the case statement did
not check for the correct long syntax.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>