sh-functions: rename fetch_all_headers
The fetch_all_headers name is not descriptive, as it actually splits the
mailbox up and processes the first header of each message. It does not
obtain every copy of a given header from the mbox. This patch renames
the function, and updates the comment so that it is clear what the
function does.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>