test-patchset: don't skip email when pre-patch build fails
Currently, whenever the pre-patched kernel fails to build, we
don't show the results for the post-patched kernel. This is faulty,
because we could easily be providing a patch which fixes the issue.
Instead, we should show the results for the failed build, but still
continue. If the post-patched kernel fails to build there is little need
to show a complex diff, but we can at least show the results for both
the pre-patched and post-patched kernel.

This patch also additionally adds --keep-going to the pre-patched kernel
in order to obtain more data for comparison.

Artem: add a commentary

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>