Fix fetch_header_or_die
The fetch_header_or_die function is suppoed to make the caller exit if the
header is not found. However, this did not happen, because the interface of
this function assumed that it is called in a subshell in order to intercept the
results. And the exiting happened in the subshell, which did not affect the
caller at all.

Fix this issue by changing the interface of fetch_header_or_die() so that the
result is written to the first parameter.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>