aiaiai: don't assume kernel tree will be checked out
This change addresses a possible issue if someone uses a bare repsitory
as their remote. We handle this by using a new sh-function git_dir()
which will return $path/.git if that is a directly, then path if that is
a directory, or will complain if it's not a valid directory. This
enables us to automatically use the correct directory without issue even
if the repository is bare.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>