sh-functions: refine the 'prefix_format' variable
First of all, the whole subject parsing thing belongs to the e-mail stuff, so
it really should be moved to the email sub-directory.

The 'prefix_format' is a variable which is supposed to explain in a short form
what is the format of the subject we expect. Currently we have a confusing
<project> part there, and this patch removes it. I do not think the result is
ideal, since it does not tell that the format is rather flexible, but still
it is a bit better. Later we can change this to something better.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>