email-lda: send a reply when chaining was incorrect
This patch is based on Jacob Keller's work. It teaches the aiaiai-email-lda
script sending rejection notifications to the patch submitter. Right now it
only sends a reply when the submitter missed the 'In-Reply-To:' e-mail header.
Later more conditions may be added.

At the moment we do not CC the 'always_cc' list, neither honor the
'reply_to_all' flag. This may be changed later, if needed.

I've added several 'program_required's too, I was too lazy to separate them,
again, sorry.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>