email-lda: honor 'always_cc' and 'reply_to_all'
When rejecting a patch we should try to honor the 'always_cc' and
'reply_to_all' per-project configuration variables. This patch does exactly
that - parses the 'To' and 'Cc' list, finds out which project the e-mail
belongs to, parses the config file and CCes the 'always_cc' list, as well as
preserves the 'To' and 'Cc' addresses if 'reply_to_all' is true.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>