test-patchset: improve build failure output
The previous patch made it acceptable for the base commit to fail. However, the
e-mail message we generate in this case contained something like:

"Failed to build the following commit for configuration ..."

which is not too readable, because it does not make it clear that this is the
base commit which fails to compile.

This patch improves the output and makes it look like this:

"Failed to build base commit ..."

This is just a bit nicer.

Additionally, when the base commit fails, and the first patch fixes it, the
resulting build diff is rather messy. This patch makes aiaiai-test-patchset
print a short explanation for the messy build diff.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>