doc: add the TODO list
Add the TODO file. I've added some of my current wishes there. This does not
mean that I am going to implement this. But may be some time, when I have time
or if users request. Or may be someone else implements one of those.

Others are welcome to add more wishes there too.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>