email-sh-functions: "export" __ini_config_get_or_die
In this project we try to mark all internal functions with leading "__". Of
course, we do this only for "shared" files like "aiaiai-email-sh-functions".
Any symbol which is not supposed to be used ouside of the shared function
starts with "__".

Thie patch removes the prefix from the "__ini_config_get_or_die" helper because
we will need it in the next patch.
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>