email-test-patchset: fix projects list generation
This patch fixes the 'list_projects()' which did not work because it was using
the 'cfg_descr' variable which was undefined. Indeed, this function is called
when the user either forgot to specify the project, or specified a non-existing
project.

This patch uses 'ini_config_get_or_die' to get project description. The whole
loop conctruct is cleaned up a bit too to make it more readable.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>