email-test-patchset: improve readability
I've noticed that this script is rather difficult to read because of many
"DocHere" usage, which break indentation. These are used when we send replies
back to the patch submitter.

This patch is an attempt to improve this. It introduces a number of helper
functions which incorporate the e-mail sending. The main script then contains
just a function calls, which improves readability.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>