Switch to dash when possible
Aiaiai is written in a portable manner, meaning that it does not use Bash
features and should work with different shells, like dash and ksh. Actually,
dash is a nice, small, and fast shell implementing only POSIX feature (or may
be mostly), and it is good to use it when it is available. This will ensure
that we do not accidentally break protability.

This patch adds a small code snippet to each script which will check if dash is
installed in the system, and if it is, "re-start" with dash.

People hacking on Aiaiai are encouradge to install the 'dash' package on their
systems.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>