Rename 'make_options' to 'kmake_opts'
The command-line option corresponding to the 'make_options' configuration knob
is called '--kmake-opts'. Let's be consistent and use the same names for both,
just to make things be more logical.

Alternatively, if 'kmake_opts' is not the best name, we could rename both to
'make_options' instead, or to something else.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>