email-sh-functions: rename project configuration variables
This is another cosmetic change which is supposed to improve the readability.
Let's distinguish between global and project parameters by using slightly
differnet variabls prefix: "cfg_" for global parameters and "pcfg_" for
per-project. Then in the contexts where we do not yet know the project, we'd
not use any of the "pcfg_" variables, and possibly avoid introducing bugs.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>