email-test-patchset: remove a useless variable
This is another cosmetic clean-up. Just trying to make this script a bit
smaller and more readable. Remove the 'msgname' variable since we only use it
once.

Also, substitute 'echo' with 'printf', since we are trying to not using echo as
a general policy, see this URL for more infor why echo may be dangeros and
printf is safe:

http://unix.stackexchange.com/questions/65803/why-is-printf-better-than-echo

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>