email-test-patchset: further improve readability
This script was confusing and difficult to follow because it used 'to', 'cc',
and 'from' variables when composing the reply. E.g., "to" means the "To:"
header of the patch under test, but when used in context of the reply, we used
it for creating the "Cc:" header. Very confusing and easy to mess things up.

This patch cleans that up. Let's leave 'to', 'cc', 'from', 'subj', and 'id'
variables to represent the _original_ "To:", "Cc:", "From:", "Subj:", and
"Id:".

For replies, introduce 'reply_to', 'reply_cc', 'reply_subj', and 'reply_id'
variables.

The code becames a lot easier to follow with this change.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>