email-sh-functions: fix e-mail addresses processing
This patch is a bug-fix. It fixes the 'strip_address()' and 'merge_addresses()'
which were for some reason removing all the blanks from the list of addresses.
The result was that "Artem Bityutskiy" became "ArtemBityutskiy".

Instead of removing all blanks, it is supposed to make multiple blanks to
become a single white-space.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>