email-lda: introduce the 'queue_series' helper
Artem: this is part of a bigger patch from Jacob, which I made to be a separate
patch. So Jacob is the author, but I am writing the commit message.

Introduce a separate function for queueing a patch series. This is a bit more
readable, and also this is a preparation for further changes.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>