email-lda: make the queued file name saner
When aiaiai-email-lda collects all patches in the series, it concatenates them
into a single file, and calls 'queue_mboxfile'. This function basically saves
a copy of the mbox in the "queue_saved" directory, and then moves it to the
queue directory.

The name of the file will be like this: date_m_of_n-counter.

However, the "m" part makes no sense at all. It is equivalent to the mumber of
the patch which came last. For example, if we have a series of 2 patches, and
path 1 came last, the name will be:

date_1_of_2-counter

Or if patch 2 came last, the name will be

date_2_of_2-counter

And this makes no sense. Just remove the "m" part from the name and make it to
be:

date_2-counter.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>