email-lda: centralise naming decision
The 'sec_name' function basically defines the name of the mbox file or a
directory for use. Currently we use the date as the prefix. Later we may change
this.

Anyway, make 'sec_name' always use the current date, isntead of getting it from
the caller. This should make things a bit easier to understand.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>