email-lda: improve readability some more
This patch adds some useful commentaries. Also, it partially makes the mbox
parameter to be non-global. Next patches will do more in this direction. This
idea is to improve code readability but making it using less global variables.
Ideally there should be none, but I am not sure I'll have time to do that.
Let's do this for mbox and probably "m" and "n". See next patches.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>