Rename 'external' to 'helpers'
I would like to move all the internal tools to a separate directory. Let's call
it 'helpers'. The 'external' directory now contains all the tools which came
from an external project. Probably this is not the best way to lay-out our
files. Let's just re-name it to helpers. Later we'll move more stuff to
'helpers'.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>