Compile helper tools if necessary
Some of our tools like 'remap-log' have to be compiled. This is not a problem
when using Aiaiai from a package, because it has the compiled version of the
tool, but this is a problem when using Aiaiai from sources.

Introduce a function which tries to compile the internal tools. This is
supposed to improve user experience with Aiaiai.

Remove the top-level Makefile since it is not really needed after this change.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>