email-test-patchset: fix pcfg_reply_to_all
Through the course of several refactors, one of the patches
accidentallly broke reply_to_all functionality. In the older code, we
did a negative check, and if reply_to_all was not set, we would then
unset $to in order to prevent the compose_email function/code from
adding any Cc addresses.

However, one of the refactors missed this negative check, assumed it was
positive, and then changed the semantics so that reply_to_all = 0
caused reply behavior rather than otherwise.

This patch corrects the given commit by making it check for 1, instead
of not 1.

Artem: amend an weird comment while on it

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Fixes: 092f2ca246 "(email-test-patchset: further improve readability)"
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>