aiaiai: don't cat the $mbox file at begining of email-testpatchset
This patch corrects an inadverdent addition of cat "$mbox" which apears
to come from commit bbf0119cbfc0. Possibly originating as some sort of
debugging code, it has essentially been lost in a rebase most likely.
However, this cat of the mbox contents is quite annoying. Possibly add
this back as a feature, since it sometimes proves useful, but generally
slows down aiaiai due to sometimes printing an exceptional amount of
data to stdout! Also fixes a typo of private.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Fixes: bbf0119cbfc0 ("email-lda: use cover letter subject if available")
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>