aiaiai-test-patchset: move --targets option down
This patch has no functional changes. It is just a clean-up which makes the
help output a tiny bit easier to follow by moving --target out of the
block of options related to checkers. Also, it changes the orger of option
handling to match the order in the help output.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>