aiaiai-checker: ignore debugging output from spatch
spatch is a great tool, but it is still not very stable and crashes sometimes,
in which case it prints a lot of debugging stuff to stderr, which we do not
want in out logs. Normal warnings go to stdout. Therefore, ignore stderr output
of spatch.

And switch to the new style options format instead of using the old style.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>