email-test-configuration: add helper program to debug configuration
This script is very useful for debugging the configuration file. It
takes a snapshot of the environment before and after parse_config and
after each parse_prj_config, and then uses diff to display the output.
This makes it very easy to see if the configuration is being read as
desired. This script has already helped debug an issue in the
shell_ini_config parser.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>