aiaiai: cleanup display of randconfig builds
This patch removes some code attempting to show the random configuration
generated for the build. It does not correctly show up in the email, so
I need to spend some time figuring out where to display the
configuration file.

Additionally, clean up some display about generating the random config
so that we don't see the errors. Also we do a make mrproper to cleanup
any generated files so that builds will be clean when we start.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>