aiaiai: test all patches when testing bisection
Instead of only testing until a patch series fails, we should test every
patch. This provides more information, and does not seriously harm the
output. To take advantage of this, we also have to modify our check for
"base commit failed but it is fixed in the first patch". We simply say
that the base commit failed and is fixed as long as the series has no
failures.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>