aiaiai: update TODO.txt to include note about randconfig
Currently there is no good way to add an attachment to the email, and so
we don't really have a way to add the random configuration as an
attachment. Update the TODO to indicate that we need a solution for this
issue. (Random configurations failures are less useful if we can't show
how to reproduce them in our response email).

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>