aiaiai-email-lda: use the References mail header
Aiaiai local mail delivery agent should be able to cope with forwarded
mails, not only git-send-email.

One popular method of forwarding a mail thread is to have it embedded in
a single mulitpart mail in rfc822 format. But this transformation
changes the In-Reply-To: headers into References: headers.

The goal of this patch is to teach aiaiai LDA to understand both
In-Reply-To and References in its mail handling. This enables mail
forwarding.

Jacob:
  Fix up variable quotes in the README example.
  Fix up style for mail headers, and fix a typo.

Signed-off-by: Robert Jarzmik <robert.jarzmik@free.fr>
Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>