aiaiai-project-update: ignore repos without a path
If the repository isn't yet checked out, don't completely fail. Future
TODO is to try and figure out a way to automatically clone the
repository. However, currently repositories are stored per-project and
multiple URLs could end up being used as remotes, which makes it
difficult to determine the right way to clone the repository. In
addition, reverse-lookup of project from path is not really easy to do.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>