aiaiai-email-sh-functions: allow patches without PATCH prefix
Not all patches are submitted with a PATCH prefix. This patch enables
the PATCH prefix to be left out, and hopefully cover more patches in
testing.

Future work should include a more robust way of determining if an email
is a patch that does not rely on the subject format, such as searching
for the diff line.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>