email: skip applied patches in autodetect project
Update the project autodetection hook to use git-patch-id to search for
patches which have already been applied. This helps ensure that patches
which have already been applied to the "branch_base" are skipped with a
different warning message. Otherwise, git-find-base will not work for
these patches, and they will be sent a more confusing message.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>