autodetect-project: limit search to nearest annotated tag
This patch limits the search for patching patch-id commits to only check
commits since the most recent tag. This works well for the Linux kernel,
as we only worry about exact resubmissions of recent work, that may have
been applied before aiaiai could test it. This may not work for all
projects, and probably some sort of configurable setting should be
created.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>