email-lda: use git-apply to more accurately find patch emails
Instead of using the subject as the sole arbiter of whether an email
contains a patch, reject all emails which cannot be recognized by
git-apply. In this way, we can locate even patches with non-standard
subject lines. Use the subject line to differentiate between patch
series and individual patches when possible.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>